requirements.txt
# =============================================================================
# DEPRECATION WARNING:
#
# The file `requirements.txt` does not influence the package dependencies and
# will not be automatically created in the next version of PyScaffold (v4.x).
#
# Please have look at the docs for better alternatives
# (`Dependency Management` section).
# =============================================================================
#
# Add your pinned requirements so that they can be easily installed with:
# pip install -r requirements.txt
# Remember to also add them in setup.cfg but unpinned.
# Example:
# numpy==1.13.3
# scipy==1.0
#
sdeint==0.2.1
pytorch==1.4.0 --install-option='-f https://download.pytorch.org/whl/torch_stable.html'
torchvision==0.5.0 --install-option='-f https://download.pytorch.org/whl/torch_stable.html'
torchtext
spacy
gpytorch
networkx
pandas
dill
revtok

